renamed GlobalConstants to Constants
